:recycle:[refactor]: pages/mypage 폴더 생성
